Fix name of options provider registration method
AddJsonSerializerSettingsProvider -> AddJsonSerializerOptionsProvider

The old name was the same as the one used by the Newtonsoft library. This is unhelpful because in some test scenarios we need to be able to call both (until such time as everything has moved over to System.Text.Json). And in any case, the thing it provides is called JsonSerializerOptions.